week-8/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.32.3
pyotp==2.9.0
qrcode==7.4.2
email-validator==2.1.0
